Revert "ci: add bootstrap script for consistent git operations"
This reverts commit 7f44e8e845ba0b236c0ea8eb178f10fc2b80e2ea.

It seems there's no way to avoid a race condition with git submodules.
The git submodule is committed at a reference that's
only available in a mozilla PR, and there's
no way to add the reference to the submodule during or before init.

IOW, I did this locally:

    cd src/app
    git config --add remote.upstream.fetch '+refs/pull/*/head:refs/remotes/upstream/pr/*'
    git fetch --all
    git checkout upstream/pr/433
    cd ../..
    git commit -m 'upgrade ...'

Then any other user, such as circleci, tries to git submodule update, and it fails.
It fails because the fetch for PRs hasn't been configured locally yet.